+++ examples/browser-sync/index.js
@@ -1,8 +1,8 @@
 /**
  * Module dependencies.
  */
-var browserSync = require('browser-sync').create()
-var proxy = require('../../index') // require('http-proxy-middleware');
+var browserSync = require('browser-sync').create();
+var proxy = require('../_proxy'); // require('http-proxy-middleware');
 
 /**
  * Configure proxy middleware
@@ -11,7 +11,7 @@ var jsonPlaceholderProxy = proxy('/users', {
   target: 'http://jsonplaceholder.typicode.com',
   changeOrigin: true, // for vhosted sites, changes host header to match to target's host
   logLevel: 'debug'
-})
+});
 
 /**
  * Add the proxy to browser-sync
@@ -23,7 +23,7 @@ browserSync.init({
     middleware: [jsonPlaceholderProxy]
   },
   startPath: '/users'
-})
+});
 
-console.log('[DEMO] Server: listening on port 3000')
-console.log('[DEMO] Opening: http://localhost:3000/users')
+console.log('[DEMO] Server: listening on port 3000');
+console.log('[DEMO] Opening: http://localhost:3000/users');